% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebooks}
\format{
An object of class \code{list} of length 2.
}
\usage{
codebook
}
\description{
Separate for the waves...
}
\details{
In particular \code{codebook$w1$all} which contains a nested list for each variable in w1 (\code{data_w}).
Can be used to quickly access question text and other stuff
}
\keyword{datasets}
