% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pe}
\alias{data_pe}
\title{Main data...}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3456 rows and 132 columns.
}
\usage{
data_pe
}
\description{
...derived from w3_pe_start_settings or w3_pe_end_settings.
Align closely with the picture of the PE evaluator and follows a
consistent nameing convention: PE_varialbe__detaill
}
\details{
Variables are of expected type.

Contains sjlabels: \code{View(data_pe)}
}
\seealso{
\code{\link[=select_attr(), get_attrs]{select_attr(), get_attrs()}}
}
\keyword{datasets}
