% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_w}
\alias{data_w}
\title{Main data...}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3456 rows and 92 columns.
}
\usage{
data_w
}
\description{
...NOT derived from w3_pe_start_settings or w3_pe_end_settings
}
\details{
Variables are of expected type but need to be cast for modeling.
However, good to get an understanding.

Contains sjlabels: \code{View(data_w)}
}
\seealso{
\code{\link[=translate_names]{translate_names()}}
}
\keyword{datasets}
