% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE.R
\name{E}
\alias{E}
\title{Expectation}
\usage{
E(..., weights = "N", keep_df = FALSE)
}
\arguments{
\item{...}{What to compute the expectation of}

\item{weights}{The name of the column containing sample weights. Set to NULL
for no weighting.}

\item{keep_df}{If TRUE, the resulting data frame is returned (which will
contain additional columns for conditioned variables).}
}
\value{
A numeric vector containing the computed expectation.
}
\description{
Expectation
}
\note{
Assumes a data frame named "dat" is defined in the global environment.
This will be modified in future versions.
}
\examples{
E(Y)
E(Y | A == 0)
E(Y | A)
E(Y | A == 0, C)
}
