% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_us_treasury.R
\name{join_us_treasury}
\alias{join_us_treasury}
\title{Join Market Yield on U.S. Treasury Securities at 2-Year Constant Maturity}
\usage{
join_us_treasury(data)
}
\arguments{
\item{data}{tibble with PE data (see load_shiller_pe)}
}
\value{
tibble with US treasury joined onto input data ordered by Date
}
\description{
Join Market Yield on U.S. Treasury Securities at 2-Year Constant Maturity
}
