% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels}
\alias{labels}
\title{Labels}
\format{
An object of class \code{list} of length 4.
}
\usage{
labels
}
\description{
In particular \code{labels$all}
}
\details{
\code{data.frame} containing from to pairs.
}
\keyword{datasets}
