% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PE.R
\docType{package}
\name{PE-package}
\alias{PE-package}
\alias{PE}
\title{Package for Priority Evaluator (PE)}
\description{
Mainly to organize data prep and distribute PE data for MDEV analysis of the experiment.
}
\examples{
data(package = "PE")
help(package = "PE")
?PE
?data_w
}
\author{
Daniel Heimgartner \href{mailto:daniel.heimgartner@ivt.baut.ethz.ch}{daniel.heimgartner@ivt.baut.ethz.ch}
}
