% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pe.R
\name{plot_pe}
\alias{plot_pe}
\title{Plot PE as a scatter plot with only lines}
\usage{
plot_pe(data)
}
\arguments{
\item{data}{tibble with columns Date and PE (see load_shiller_pe)}
}
\value{
plotly object
}
\description{
Plot PE as a scatter plot with only lines
}
