% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_attr.R
\name{select_attr}
\alias{select_attr}
\title{Select but for attrs}
\usage{
select_attr(df, var, attr = "group")
}
\arguments{
\item{df}{\code{data.frame} where individual columns have an \code{attr} attribute}

\item{var}{to select (character). Not vectorized...}

\item{attr}{which attribute to target (defaults to \code{"group"})}
}
\value{
Subset of \code{data.frame} wherre attr == var
}
\description{
What \code{dplyr::select()} does for columns, \code{select_attr()} does for attributes
(not as good though)
}
\examples{
select_attr(data_pe, "co2_final")
}
