% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_names.R
\name{translate_names}
\alias{translate_names}
\title{Translate names of \code{data_w}}
\usage{
translate_names(nm, vnt = variable_names_table, current = c("from", "to"))
}
\arguments{
\item{nm}{\code{names(data_w)}}

\item{vnt}{from to mapping}

\item{current}{Does \code{nm} currently correspond to \code{from} or \code{to} in \code{vnt}?
Defaults to \code{"from"}. Intended to translate back if necessary...}
}
\value{
character vector with translated names
}
\description{
Translate names of \code{data_w}
}
\examples{
print(names(data_w))
names(data_w) <- translate_names(names(data_w))
print(names(data_w))
}
